Uploading important files for To-do list app
Important to note: These are only a selection of the files uploaded (primarily the code written to create the webpages) and uses Bootstrap as the main frontend toolkit. If you were to run the program or open it (even on this computer), it wouldn't compile or run due to only a selection (most valuable files in my opinion) that have been uploaded to show what has been done. 

IMPORTANT: This first project has been completed through a tutorial, I added some new features myself but about 95% of the project is through the tutorial. I am using this as a base to start my own projects, although I may need a few more till I really get started. Thank you